ai-agents/pharma-rag-agent/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-text-splitters
langchain-chroma
langchain-huggingface
langchain-groq
sentence-transformers
pypdf
python-multipart
